// frontend/src/types/enums.ts
export enum RoleType {
  SUPER_ADMIN = 'SUPER_ADMIN',
  ADMIN = 'ADMIN',
  EDITOR = 'EDITOR',
  VIEWER = 'VIEWER',
}
export enum FieldType {
  TEXT = 'TEXT',
  TEXTAREA = 'TEXTAREA',
  NUMBER = 'NUMBER',
  EMAIL = 'EMAIL',
  PHONE = 'PHONE',
  DROPDOWN = 'DROPDOWN',
  CHECKBOX = 'CHECKBOX',
  RADIO = 'RADIO',
  DATE = 'DATE',
  TIME = 'TIME',
  RATE = 'RATE',
  HEADER = 'HEADER',
  FULLNAME = 'FULLNAME',
  ADDRESS = 'ADDRESS',
  PARAGRAPH = 'PARAGRAPH',
  SUBMIT = 'SUBMIT',
  DIVIDER = 'DIVIDER',
  SECTION_COLLAPSE = 'SECTION_COLLAPSE',
  PAGE_BREAK = 'PAGE_BREAK',
  GROUP = 'GROUP',
  MATRIX = 'MATRIX',
  TABLE = 'TABLE',
  FILE = 'FILE',
}
export enum FormStatus {
  DRAFT = 'DRAFT',
  PUBLISHED = 'PUBLISHED',
  ARCHIVED = 'ARCHIVED',
}